# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Local & Experiences API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me/profile": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Put Profile",
        "operationId": "put_profile_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Profile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/password": {
      "put": {
        "tags": [
          "me"
        ],
        "summary": "Change Password",
        "operationId": "change_password_me_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Delete Me",
        "operationId": "delete_me_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/context/import": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Import Context",
        "description": "A past itinerary or trip notes in your own words (\"Goa last winter: Old Goa churches,\nFort Aguada, beach shacks; skipped the casinos\") -> what you liked and avoided.",
        "operationId": "import_context_me_context_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/context/{tag}": {
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Forget Tag",
        "operationId": "forget_tag_me_context__tag__delete",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "local-food",
                "street-food",
                "fine-dining",
                "vegetarian",
                "sweets",
                "chai",
                "heritage",
                "history",
                "architecture",
                "museum",
                "spiritual",
                "craft",
                "art",
                "music",
                "dance",
                "performance",
                "workshop",
                "hands-on",
                "shopping",
                "market",
                "textiles",
                "jewellery",
                "nature",
                "wildlife",
                "beach",
                "sunset",
                "sunrise",
                "viewpoint",
                "photography",
                "adventure",
                "active",
                "wellness",
                "relaxed",
                "nightlife",
                "evening",
                "social",
                "romantic",
                "family",
                "kids",
                "learning",
                "community",
                "walking-tour",
                "hidden-gem",
                "iconic"
              ],
              "type": "string",
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Companion": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 110.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "interests": {
            "items": {
              "type": "string",
              "enum": [
                "local-food",
                "street-food",
                "fine-dining",
                "vegetarian",
                "sweets",
                "chai",
                "heritage",
                "history",
                "architecture",
                "museum",
                "spiritual",
                "craft",
                "art",
                "music",
                "dance",
                "performance",
                "workshop",
                "hands-on",
                "shopping",
                "market",
                "textiles",
                "jewellery",
                "nature",
                "wildlife",
                "beach",
                "sunset",
                "sunrise",
                "viewpoint",
                "photography",
                "adventure",
                "active",
                "wellness",
                "relaxed",
                "nightlife",
                "evening",
                "social",
                "romantic",
                "family",
                "kids",
                "learning",
                "community",
                "walking-tour",
                "hidden-gem",
                "iconic"
              ]
            },
            "type": "array",
            "title": "Interests",
            "default": []
          },
          "dislikes": {
            "items": {
              "type": "string",
              "enum": [
                "local-food",
                "street-food",
                "fine-dining",
                "vegetarian",
                "sweets",
                "chai",
                "heritage",
                "history",
                "architecture",
                "museum",
                "spiritual",
                "craft",
                "art",
                "music",
                "dance",
                "performance",
                "workshop",
                "hands-on",
                "shopping",
                "market",
                "textiles",
                "jewellery",
                "nature",
                "wildlife",
                "beach",
                "sunset",
                "sunrise",
                "viewpoint",
                "photography",
                "adventure",
                "active",
                "wellness",
                "relaxed",
                "nightlife",
                "evening",
                "social",
                "romantic",
                "family",
                "kids",
                "learning",
                "community",
                "walking-tour",
                "hidden-gem",
                "iconic"
              ]
            },
            "type": "array",
            "title": "Dislikes",
            "default": []
          },
          "accessibility": {
            "items": {
              "type": "string",
              "enum": [
                "wheelchair",
                "step_free",
                "seating",
                "quiet"
              ]
            },
            "type": "array",
            "title": "Accessibility",
            "default": []
          },
          "diet": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "vegetarian",
                  "non_vegetarian",
                  "vegan",
                  "jain"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Diet"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Companion",
        "description": "A person the user often travels with; reused when planning trips (P3)."
      },
...
      "ContextEntry": {
        "properties": {
          "tag": {
            "type": "string",
            "enum": [
              "local-food",
              "street-food",
              "fine-dining",
              "vegetarian",
              "sweets",
              "chai",
              "heritage",
              "history",
              "architecture",
              "museum",
              "spiritual",
              "craft",
              "art",
              "music",
              "dance",
              "performance",
              "workshop",
              "hands-on",
              "shopping",
              "market",
              "textiles",
              "jewellery",
              "nature",
              "wildlife",
              "beach",
              "sunset",
              "sunrise",
              "viewpoint",
              "photography",
              "adventure",
              "active",
              "wellness",
              "relaxed",
              "nightlife",
              "evening",
              "social",
              "romantic",
              "family",
              "kids",
              "learning",
              "community",
              "walking-tour",
              "hidden-gem",
              "iconic"
            ],
            "title": "Tag"
          },
          "weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Weight"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated"
          }
        },
        "type": "object",
        "required": [
          "tag",
          "weight",
          "source"
        ],
        "title": "ContextEntry"
      },
...
      "ContextImport": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ContextImport",
        "description": "A past itinerary or trip notes, in the traveler's words."
      },
      "DeleteAccount": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 128,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "DeleteAccount"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChange": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange"
      },
...
      "Profile": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Display Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 110.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "home_city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Home City"
          },
          "interests": {
            "items": {
              "type": "string",
              "enum": [
                "local-food",
                "street-food",
                "fine-dining",
                "vegetarian",
                "sweets",
                "chai",
                "heritage",
                "history",
                "architecture",
                "museum",
                "spiritual",
                "craft",
                "art",
                "music",
                "dance",
                "performance",
                "workshop",
                "hands-on",
                "shopping",
                "market",
                "textiles",
                "jewellery",
                "nature",
                "wildlife",
                "beach",
                "sunset",
                "sunrise",
                "viewpoint",
                "photography",
                "adventure",
                "active",
                "wellness",
                "relaxed",
                "nightlife",
                "evening",
                "social",
                "romantic",
                "family",
                "kids",
                "learning",
                "community",
                "walking-tour",
                "hidden-gem",
                "iconic"
              ]
            },
            "type": "array",
            "title": "Interests",
            "default": []
          },
          "dislikes": {
            "items": {
              "type": "string",
              "enum": [
                "local-food",
                "street-food",
                "fine-dining",
                "vegetarian",
                "sweets",
                "chai",
                "heritage",
                "history",
                "architecture",
                "museum",
                "spiritual",
                "craft",
                "art",
                "music",
                "dance",
                "performance",
                "workshop",
                "hands-on",
                "shopping",
                "market",
                "textiles",
                "jewellery",
                "nature",
                "wildlife",
                "beach",
                "sunset",
                "sunrise",
                "viewpoint",
                "photography",
                "adventure",
                "active",
                "wellness",
                "relaxed",
                "nightlife",
                "evening",
                "social",
                "romantic",
                "family",
                "kids",
                "learning",
                "community",
                "walking-tour",
                "hidden-gem",
                "iconic"
              ]
            },
            "type": "array",
            "title": "Dislikes",
            "default": []
          },
          "accessibility": {
            "items": {
              "type": "string",
              "enum": [
                "wheelchair",
                "step_free",
                "seating",
                "quiet"
              ]
            },
            "type": "array",
            "title": "Accessibility",
            "default": []
          },
          "walking_limit_km": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Walking Limit Km"
          },
          "needs_rest_breaks": {
            "type": "boolean",
            "title": "Needs Rest Breaks",
            "default": false
          },
          "diet": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "vegetarian",
                  "non_vegetarian",
                  "vegan",
                  "jain"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Diet"
          },
          "pace": {
            "type": "string",
            "enum": [
              "relaxed",
              "normal",
              "packed"
            ],
            "title": "Pace",
            "default": "normal"
          },
          "budget_style": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "budget",
                  "mid",
                  "premium"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Style"
          },
          "transport": {
            "items": {
              "type": "string",
              "enum": [
                "walk",
                "auto",
                "bus",
                "car"
              ]
            },
            "type": "array",
            "title": "Transport",
            "default": []
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Languages",
            "default": []
          },
          "companions": {
            "items": {
              "$ref": "#/components/schemas/Companion"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Companions",
            "default": []
          },
          "avoid_crowds": {
            "type": "boolean",
            "title": "Avoid Crowds",
            "default": false
          },
          "hidden_gems": {
            "type": "boolean",
            "title": "Hidden Gems",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "Profile",
        "description": "Everything used to personalise recommendations. All optional except the name.\nSensitive fields (age, accessibility, diet) are owner-only: never shown to admins/providers."
      },
      "ProfileContext": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ContextEntry"
            },
            "type": "array",
            "title": "Entries",
            "default": []
          },
          "from_trips": {
            "items": {
              "$ref": "#/components/schemas/ContextEntry"
            },
            "type": "array",
            "title": "From Trips",
            "default": []
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          }
        },
        "type": "object",
        "title": "ProfileContext",
        "description": "What the planner has learned about the traveler, visible and correctable (doc §12.2)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}